In Progress

+++ TestSutraUsers/src/main/java/com/tathvatech/unit/common/TestableEntity.java
@@ -6,6 +6,6 @@ public interface TestableEntity
 {
 	public OID getOID();
 
-	public int getPk();
+	public long getPk();
 
 }
